Improve connection handling with multiple addresses
The problem was that connections which failed to connect were added to the active connections.
This caused issues with https://github.com/syncthing/syncthing-java/pull/14 as often one of the failed connections was used.

Now, connecting over the internet/ relays works "reliable" if the Syncthing Server thinks
that Syncthing Lite is not connected (shows "not connected" in the Web UI). To get
this situation, one has to restart the server.